% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_query.R
\name{build_query}
\alias{build_query}
\title{Build a query}
\usage{
build_query(identify, filter, location, select = NULL, profile = NULL)
}
\description{
This is ported from \code{galah/utilities_internal.R}, where it is an internal
function. Designed to take standardised \code{galah} objects and turn them into
queries
}
