% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_solr.R
\name{parse_solr}
\alias{parse_solr}
\title{parse a query into \code{solr} syntax}
\usage{
parse_solr(df)
}
\arguments{
\item{df}{A \code{tibble} containing columns \code{variable}, \code{logical} and \code{value}}
}
\value{
A vector of valid solr queries
}
\description{
Many living atlases use \code{solr} as their indexing engine. This function takes
a \code{tibble} of specified structure, and parses it to return a valid \code{solr}
query.
}
