% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_orgs.R
\name{show_orgs}
\alias{show_orgs}
\title{Show orgs}
\usage{
show_orgs()
}
\description{
Returns a \code{tibble} of supported biodiversity organisations with working APIs.
Note that not all GBIF nodes are supported currently by \code{honeybee}
}
